refactor: extract static method
