+++ ProtoCredentials/OpenID4VC-Prototype/Presentation/Console/CvFlowPresentation.cs
@@ -5,26 +5,53 @@ using OpenID4VC_Prototype.Domain.Models;
 using Serilog;
 
 namespace OpenID4VC_Prototype.Presentation.Console;
+
 public static class CvFlowPresentation
 {
-    public static void IssueCV(DecentralizedIdentifier issuer, IIssuerService issuerService, DecentralizedIdentifier holder, out VCDto credential)
+    public static void IssueVC(DecentralizedIdentifier issuer, IIssuerService issuerService, DecentralizedIdentifier holder, IVerifierService verifierService)
     {
+        // Issuing a verifiable credential
+        WriteTitle("Issuing verifiable credential");
+        var credential = new VCDto();
         try
         {
             var issuerDto = issuer.Adapt<DIdDto>();
             credential = issuerService.IssueCredential(issuerDto, holder.DId);
 
-            Log.Information($"Issued Credential: {credential.CredentialType} for {credential.HolderDId}");
+            System.Console.WriteLine($"Issued Credential: {credential.CredentialType} for {credential.HolderDId}");
         }
         catch (ArgumentException ex)
         {
             Log.Warning(ex, $"Issuing credential failed: {ex.Message}");
-            credential = new VCDto();
         }
         catch (Exception ex)
         {
             Log.Fatal(ex, $"Unexpected error: {ex.Message}");
-            credential = new VCDto();
         }
+
+        // Verifier validates the credential
+        WriteTitle("Verifier validates the credential");
+        try
+        {
+            var validationResult = verifierService.ValidateCredential(credential, issuer.PublicKey);
+
+            Log.Information(validationResult.IsValid
+                ? "Credential is valid!"
+                : $"Verification failed: {validationResult.ErrorMessage}");
+        }
+        catch (ArgumentException ex)
+        {
+            Log.Warning(ex, $"Validation error: {ex.Message}");
+        }
+        catch (Exception ex)
+        {
+            Log.Fatal(ex, $"Unexpected error: {ex.Message}");
+        }
+    }
+
+    static void WriteTitle(string title)
+    {
+        System.Console.WriteLine();
+        System.Console.WriteLine("***" + title.ToUpper());
     }
 }
