+++ ProtoCredentials/OpenID4VC-Prototype/Program.cs
@@ -8,6 +8,7 @@ using OpenID4VC_Prototype.Application.Services;
 using OpenID4VC_Prototype.Domain.Interfaces;
 using OpenID4VC_Prototype.Domain.Services;
 using OpenID4VC_Prototype.Infrastructure.Configurations;
+using OpenID4VC_Prototype.Presentation.Console;
 using Serilog;
 using  didConfig = OpenID4VC_Prototype.Infrastructure.Configurations.DIdConfig;
 
@@ -45,50 +46,6 @@ Console.WriteLine($"Issuer DID: {issuer.DId}");
 Console.WriteLine($"Holder DID: {holder.DId}");
 Console.WriteLine($"Verifier DID: {verifier.DId}");
 
-// Issuing a verifiable credential
-WriteTitle("Issuing verifiable credential");
-var credential = new VCDto();
-try
-{
-    var issuerDto = issuer.Adapt<DIdDto>();
-    credential = issuerService.IssueCredential(issuerDto, holder.DId);
-
-    Console.WriteLine($"Issued Credential: {credential.CredentialType} for {credential.HolderDId}");
-}
-catch (ArgumentException ex)
-{
-    Log.Warning(ex, $"Issuing credential failed: {ex.Message}");
-}
-catch (Exception ex)
-{
-    Log.Fatal(ex, $"Unexpected error: {ex.Message}");
-}
-
-// Verifier validates the credential
-WriteTitle("Verifier validates the credential");
-try
-{
-    var validationResult = verifierService.ValidateCredential(credential, issuer.PublicKey);
-
-    Log.Information(validationResult.IsValid
-        ? "Credential is valid!"
-        : $"Verification failed: {validationResult.ErrorMessage}");
-}
-catch (ArgumentException ex)
-{
-    Log.Warning(ex, $"Validation error: {ex.Message}");
-}
-catch (Exception ex)
-{
-    Log.Fatal(ex, $"Unexpected error: {ex.Message}");
-}
+CvFlowPresentation.IssueVC(issuer, issuerService, holder, verifierService);
 
 Log.CloseAndFlush();
-
-return;
-
-static void WriteTitle(string title)
-{
-    Console.WriteLine();
-    Console.WriteLine("***" + title.ToUpper());
-}
